refactor: split types among module categories creating room to grow
