+++ packages/caravan-psbt/src/psbtv2/index.ts
@@ -1,0 +1,3 @@
+export { getPsbtVersionNumber } from "./functions";
+export { PsbtV2 } from "./psbtv2";
+export { PsbtV2Maps } from "./psbtv2maps";
