+++ packages/caravan-psbt/src/psbtv2/psbtv2.test.ts
@@ -1,4 +1,4 @@
-import { PsbtV2, getPsbtVersionNumber } from "./psbtv2";
+import { PsbtV2, getPsbtVersionNumber } from "./";
 import { test } from "@jest/globals";
 
 const BIP_370_VECTORS_INVALID_PSBT = [
@@ -729,7 +729,7 @@ describe("PsbtV2", () => {
           expect(t).toThrow();
         }
       }
-    }
+    },
   );
 
   test.each(BIP_370_VECTORS_VALID_PSBT)(
@@ -746,13 +746,13 @@ describe("PsbtV2", () => {
           expect(psbt).toBe(vect[key]);
         }
       }
-    }
+    },
   );
 
   test.each(
     BIP_370_VECTORS_VALID_PSBT.filter(
-      (vect) => vect.inputs !== undefined && vect.outputs !== undefined
-    )
+      (vect) => vect.inputs !== undefined && vect.outputs !== undefined,
+    ),
   )("Returns proper input and output counts. $case", (vect) => {
     const psbt = new PsbtV2(vect.hex);
     expect(psbt.PSBT_GLOBAL_INPUT_COUNT).toBe(vect.inputs);
@@ -764,13 +764,13 @@ describe("PsbtV2", () => {
     (vect) => {
       const psbt = new PsbtV2(vect.hex);
       expect(psbt.PSBT_IN_SEQUENCE).toEqual(vect.sequences);
-    }
+    },
   );
 
   test.each(
     BIP_370_VECTORS_VALID_PSBT.filter(
-      (vect) => vect.heightLocks || vect.timeLocks
-    )
+      (vect) => vect.heightLocks || vect.timeLocks,
+    ),
   )("Returns input locktime fields. $case", (vect) => {
     if (vect.heightLocks) {
       const psbt = new PsbtV2(vect.hex);
@@ -787,7 +787,7 @@ describe("PsbtV2", () => {
     (vect) => {
       const psbt = new PsbtV2(vect.hex);
       expect(psbt.PSBT_GLOBAL_TX_MODIFIABLE).toEqual(vect.modifiable);
-    }
+    },
   );
 
   it("Returns all PSBTv2 specific fields", () => {
@@ -799,11 +799,11 @@ describe("PsbtV2", () => {
     const sequences = psbt.PSBT_IN_SEQUENCE.filter((el) => el !== null);
     expect(sequences.length).not.toBe(0);
     const timeLocks = psbt.PSBT_IN_REQUIRED_TIME_LOCKTIME.filter(
-      (el) => el !== null
+      (el) => el !== null,
     );
     expect(timeLocks.length).not.toBe(0);
     const heightLocks = psbt.PSBT_IN_REQUIRED_HEIGHT_LOCKTIME.filter(
-      (el) => el !== null
+      (el) => el !== null,
     );
     expect(heightLocks.length).not.toBe(0);
   });
@@ -927,14 +927,14 @@ describe("PsbtV2.FromV0", () => {
           expect(t).toThrow();
         }
       }
-    }
+    },
   );
 
   it("Throws with valid psbtv0 with PSBT_GLOBAL_TX_VERSION=1 when disallowing txnv1", () => {
     const t = () =>
       PsbtV2.FromV0(
         "cHNidP8BAMUBAAAAA4RSZmhtXSRz+wmYLHLaDW1msFfD4TputL/aMEB27+dlAQAAAAD/////KgI+xaBWgfS8tWueRYhPYlqWZY4doW+ALhAuMaganq4BAAAAAP////9ErmEIocbg7uZe38fpG3ICYmN2nLh3FKmd1F24+8FD8gAAAAAA/////wIGcwQAAAAAABepFOO6EVG3Xv+/etxGc8g8j+7D3cNnh28dAAAAAAAAF6kUw01jpnIIZgcEkKjLJExr3Hzi+hOHAAAAAAABAPcCAAAAAAEBSckS0OXkb275MwOMf7fh1mXbmuVrZ/pX/kw0dqlc+VQAAAAAFxYAFADi94+YelpEk88GKZTb3knQQKki/v///wJjFBgAAAAAABepFMerbRAxgKSBgYR9NXMuk+DOmrBzh6CGAQAAAAAAF6kUhHkHLVpVDuCQC1r35wr1dVJ6h52HAkcwRAIgL1OHUuQItIF+d1HvJD7uZ9IkLKIGHo5snyKHMkfxCo0CIFtGIjFO/XM/EvxlV7wvMj/yy8FgStl6NRgH4b6Ah1vIASEC6SM19uyxhi8O6guZKX8hvbm+uaHo9BETeI9a3TBsqfzumxgAAQRHUiECqFE9mTGJbV06/IBjFI23XYhR/R/EGxCYuipqdm21Y9QhA5ON0Jvz3Snd9B8mSFisz6QLMwyY4O0nyvd3NPrAATm6Uq4iBgKoUT2ZMYltXTr8gGMUjbddiFH9H8QbEJi6Kmp2bbVj1Bj1fsZdLQAAgAEAAIBkAACAAAAAAAAAAAAiBgOTjdCb890p3fQfJkhYrM+kCzMMmODtJ8r3dzT6wAE5uhgAAAABLQAAgAEAAIBkAACAAAAAAAAAAAAAAQD3AgAAAAABAQF0Xh2qKMFwXb9z7dGD5e+RrQkY2XrT4uwsabVICG9NAAAAABcWABQrC1IrqH2xZGiYEYhgRJ/LLGna4/7///8CMpZCAAAAAAAXqRQPiU9+O3C4dB+DDgZrbvUIqfdHnYeghgEAAAAAABepFIR5By1aVQ7gkAta9+cK9XVSeoedhwJHMEQCIC3Ih+XWI72XSWgoXpyBZc+p+s2UPK8PhHLnrO9jL7lDAiBcYENAYeak5FNg07PJAanB3RSLON1sliPNj6JndYfmMgEhAjZlOGkv+5Yi51oF3CAE2F76DrwnuZlh5pTYj57eK1fK5JsYAAEER1IhAqhRPZkxiW1dOvyAYxSNt12IUf0fxBsQmLoqanZttWPUIQOTjdCb890p3fQfJkhYrM+kCzMMmODtJ8r3dzT6wAE5ulKuIgYCqFE9mTGJbV06/IBjFI23XYhR/R/EGxCYuipqdm21Y9QY9X7GXS0AAIABAACAZAAAgAAAAAAAAAAAIgYDk43Qm/PdKd30HyZIWKzPpAszDJjg7SfK93c0+sABOboYAAAAAS0AAIABAACAZAAAgAAAAAAAAAAAAAEA9wIAAAAAAQHl1qD/xfg4epDEY79hSuU2CbcpiMRK/GpXfyJma8lxpwAAAAAXFgAUKDhkidFbHN39JFtQa4/y2QmxjTb+////AqCGAQAAAAAAF6kUhHkHLVpVDuCQC1r35wr1dVJ6h52Hhs4YBQAAAAAXqRTS+wqJWOVdTGw/9Y+XD9u6MAbsB4cCRzBEAiAHpxhuavuT3nSbOpBdHHQ39HD5cJXqQQU4tqwz0VqUeAIgWmYRjH3C4U1zJaEi6wAh9U4dvV37j9VrJT+jeCcWrz0BIQP1lRzMzwCWTVTu+ngoCuCD4PDwzGOC/Sez+/3+2o3Sx7KbGAABBEdSIQKoUT2ZMYltXTr8gGMUjbddiFH9H8QbEJi6Kmp2bbVj1CEDk43Qm/PdKd30HyZIWKzPpAszDJjg7SfK93c0+sABObpSriIGAqhRPZkxiW1dOvyAYxSNt12IUf0fxBsQmLoqanZttWPUGPV+xl0tAACAAQAAgGQAAIAAAAAAAAAAACIGA5ON0Jvz3Snd9B8mSFisz6QLMwyY4O0nyvd3NPrAATm6GAAAAAEtAACAAQAAgGQAAIAAAAAAAAAAAAAAAQBHUiECGgSXRxIDRfqQF/tC2P89T7HS70yAVGhyxdpRO6vVFYUhA6AAld9INn7SHlxu3VCvQ1IxG/Bg6xAEJct69DMaoarQUq4iAgIaBJdHEgNF+pAX+0LY/z1PsdLvTIBUaHLF2lE7q9UVhRgAAAABLQAAgAEAAIBkAACAAQAAAAAAAAAiAgOgAJXfSDZ+0h5cbt1Qr0NSMRvwYOsQBCXLevQzGqGq0Bj1fsZdLQAAgAEAAIBkAACAAQAAAAAAAAAA",
-        false
+        false,
       );
     expect(t).toThrow();
   });
@@ -948,13 +948,13 @@ describe("PsbtV2.FromV0", () => {
           expect(t).not.toThrow();
         }
       }
-    }
+    },
   );
 
   test.each(
     BIP_174_VECTORS_VALID_PSBT.filter(
-      (vect) => vect.inputs !== undefined && vect.outputs !== undefined
-    )
+      (vect) => vect.inputs !== undefined && vect.outputs !== undefined,
+    ),
   )("Returns proper input and output counts. $case", (vect) => {
     const psbt = PsbtV2.FromV0(vect.hex, true);
     expect(psbt.PSBT_GLOBAL_INPUT_COUNT).toBe(vect.inputs);
@@ -962,7 +962,7 @@ describe("PsbtV2.FromV0", () => {
   });
 
   test.each(
-    BIP_174_VECTORS_VALID_PSBT.filter((vect) => vect.partialSigs !== undefined)
+    BIP_174_VECTORS_VALID_PSBT.filter((vect) => vect.partialSigs !== undefined),
   )("Sets partialSigs when they are present. $case", (vect) => {
     const psbt = PsbtV2.FromV0(vect.hex, true);
     const partialSigs = psbt.PSBT_IN_PARTIAL_SIG;
@@ -1006,7 +1006,7 @@ describe("getPsbtVersionNumber", () => {
     (vect) => {
       const version = getPsbtVersionNumber(vect.psbt);
       expect(version).toEqual(vect.expected);
-    }
+    },
   );
 });
 
@@ -1028,14 +1028,14 @@ describe("PsbtV2.addPartialSig", () => {
 
     psbt.addInput({ previousTxId: Buffer.from([0x00]), outputIndex: 0 });
     expect(() => addSig(0)).toThrow(
-      "PsbtV2.addPartialSig() missing argument pubkey"
+      "PsbtV2.addPartialSig() missing argument pubkey",
     );
     expect(() => addSig(0, Buffer.from([0x00]))).toThrow(
-      "PsbtV2.addPartialSig() missing argument sig"
+      "PsbtV2.addPartialSig() missing argument sig",
     );
     addSig(0, Buffer.from([0x00]), Buffer.from([0x00]));
     expect(() => addSig(0, Buffer.from([0x00]), Buffer.from([0x00]))).toThrow(
-      "PsbtV2 already has a signature for this input with this pubkey"
+      "PsbtV2 already has a signature for this input with this pubkey",
     );
   });
 
@@ -1085,7 +1085,7 @@ describe("PsbtV2.removePartialSig", () => {
 
     psbt.addInput({ previousTxId: Buffer.from([0x00]), outputIndex: 0 });
     expect(() => removeSig(0, Buffer.from([0x00]))).toThrow(
-      "PsbtV2 input has no signature from pubkey 00"
+      "PsbtV2 input has no signature from pubkey 00",
     );
   });
 
@@ -1172,7 +1172,7 @@ describe("PsbtV2.handleSighashType (private)", () => {
     (vect) => {
       psbt.handleSighashType(vect.value);
       expect(psbt.PSBT_GLOBAL_TX_MODIFIABLE).toEqual(vect.expectedModifiable);
-    }
+    },
   );
 });
 
@@ -1234,8 +1234,8 @@ describe("PsbtV2.isModifiable (private)", () => {
       expect(psbt.isModifiable(["OUTPUTS"])).toBe(vect.outputsExpected);
       expect(psbt.isModifiable(["OUTPUTS", "INPUTS"])).toBe(vect.bothExpected);
       expect(psbt.isModifiable(["SIGHASH_SINGLE"])).toBe(
-        vect.sighashSingleExpected
+        vect.sighashSingleExpected,
       );
-    }
+    },
   );
 });
