+++ packages/caravan-psbt/src/psbtv2/values.ts
@@ -1,0 +1,4 @@
+export { PSBT_MAGIC_BYTES } from "../psbt";
+export const PSBT_MAP_SEPARATOR = Buffer.from([0x00]);
+export const BIP_32_NODE_REGEX = /(\/[0-9]+'?)/gi;
+export const BIP_32_HARDENING_OFFSET = 0x80000000;
